% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOFShiny.R
\name{GOFShiny}
\alias{GOFShiny}
\title{Launch 'GOFShiny' Interface}
\usage{
GOFShiny()
}
\value{
Nothing
}
\description{
GOFShiny() loads interactive user interface built using R 'shiny'.
}
\details{
Start GOFShiny

The interactive user interface is to provide an easy way for goodness of fit test analysis.
}
\examples{
if(interactive()){
library(rmarkdown)
GOFShiny()
}
}
\keyword{GOFShiny}
